requirements.txt
numpy==2.4.1
pandas==2.3.3

TA_Lib==0.6.8
arrow==1.4.0
bokeh==3.8.1
psycopg2-binary==2.9.9
requests==2.32.5
Logbook==1.9.2
SQLAlchemy==2.0.45
tornado==6.5.4
tqdm==4.67.1
urllib3==2.6.3
easytrader==0.23.7
beautifulsoup4==4.14.3
pycryptodome==3.23.0
python_dateutil==2.9.0.post0